scripts/requirements.txt
questionary>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
